Update redshift-profile.md
We spend some decent amount going the source code to find out that the redshift connection handler inherits from the postgres connection handler, and as such uses the psycopg2 library to connect to redshift and thus this will work as well. We verified to connection towards our redshift with both require (and not setting the sslrootcert) and verify-full (setting the redshift certs both on the default path, as well as a different path).