Update package-management.md
There is a missing `'` within the docs when referencing utilizing environment variables for private packages. 

I also noticed that this method required [percent encoding](https://en.wikipedia.org/wiki/Percent-encoding) the `@` sign when referencing email for the `GIT_CREDENTIALS` since it will fail if not escaped properly, might be worth to note in the documentation.