Add developer mode and long paths for Windows 10
dbt deps by default will leverage symlinks and Windows 10 will report them as being supported, but unless the user is in Developer Mode or an Administrator, they do not have the privileges to create symlinks.  

It possible to have files with a pathname+filename that exceeds 260 characters which is the default limit without enabling Long Paths -- this most often would affect tests where the compiled SQL filename could be quite long AND deep in the folder hierarchy.