Add semicolon to timezone example code
# Detailed description

The code example showing how to alter session in Snowflake to change the timezone is missing a semicolon. 

# Problem that this commit solves

If the semicolon is not included, `dbt run` will throw a syntax error because it immediately jump to and include the main `CREATE` statement of the model. Adding the semicolon keeps the 2 statements separate and works as intended.

# Related issue

Resolves: https://github.com/dbt-labs/docs.getdbt.com/issues/721#issuecomment-874357312