Added note about select and selectors in snapshot
Added the note below, however it does need technical confirmation as I don't know if this is correct. This is based on a zendesk ticket, https://dbtcloud.zendesk.com/agent/tickets/19492, where the user was trying to run a specific snapshot using a selector. 

Note: To run a specific snapshot, you do not need to use 'select' with a 'selector'. For example, to run a specific snapshot, you can use write: dbt snapshot --select tag:tag_name