Enhance explanation of transient table config choice
This change corrects the information about transient tables and Time Travel and also now mentions the impact on Fail-safe. I've attempted to make the information about the trade-offs of using transient tables clearer, and to also make the info about the default dbt setting more prominent as this was easily missed by new users.