Update snapshots.md/hard deletes(opt-in)
Propose clarifying the sentence "For this configuration to work, the configured updated_at column must be of timestamp type. Otherwise, queries will fail due to mixing data types.". 

At a glance it's easy to mistake this to mean that the `check_cols` strategy is not supported, which is not the case.