Update homebrew.md
I'm not an expert in the ways of Python package installations and so this may need some revisions to actually speak truth, but here's a whack at adding some hints for those of us who faced grief in our latest installation. In my case, the upgrade to 1.0 highlighted these issues.