Update snowflake-configs.md
In researching the use of time travel with dbt, I found the documentation to be mis-leading and thought I needed to change my dbt configuration. The comment that transient tables do not participate in time travel is not correct. Per Snowflake's documentation, Transient tables persist until they are dropped and the time travel default is 1 day and transient tables do not participate in fail safe.