clarify Snowflake account location
Because the url of Snowflake, where the account information can be easily found, changes based on which version of Snowflake the user is on (Classic Console vs AppUI/Snowsite), it can be confusing to identify the location of the account information. I added a distinction for each console to provide more clarity.