Add full example command line to run-operation
Properly formatting the YAML string seems easy from the usage, but without the ' on each side, it doesn't work and I had to do a lot of trial and error to research it to get it just right.  Also, I believe that the {args} in the usage should be surrounded by ' ', as they seem to be required to make it work.