augment the docs readme
made additions to readme that would explain what language is used for docs, how to name files, how to use images, folder structure, callouts, etc. 

there are some open questions for discussion that need to be 🔨 gaveled:

- do we have a naming convention for files? will they diff depending on the type of file (concept, reference, procedural)? Example content below from replicated:
`The file name should always start with the feature type (such as dbt-cloud, ide, or job-scheduler). Depending on the content type, it typically also includes a secondary descriptor and a verb. Verbs are used when you are creating a task topic.

Because we author content using Markdown, you must add the `.md` the file extension to the file name.

If you are adding a new page to an existing feature category, follow the existing naming convention for that category.

**Example: Concept file**

`snapshots-backup-hooks.md`

**Example: Reference file**

`snapshots-backup-hooks.md`

**Example: Task file**

`releases-creating-customer.md`

**Example: Tutorial file**

`tutorial-ha-cluster-deploying.md`
`

- what is the difference between alt text and lightbox src for images? Replicated use `![alt text](images/<image-name>.png)` and mention Alt text is used with image syntax, is parsed by screen readers to support accessibility. With *alt text*, the user will hear the alt text spoken out loud by the screen reader/