Add useful link to config docs
The documentation for the config function isn't that useful when it comes to describing the valid arguments. I have to hunt for the model configuration page each time to find the valid arguments. This issue is compounded by the fact that there will be no error if an *invalid* argument is supplied.