fixing source.yml code snippet
When clicking the copy button in the top right of the sources.yml code snippet, it is copying the actual .md hyperlink information, which makes the code unusable for folks. I also removed the asterisk (*) after schema, because if it is left in it will create a .yml error in the file.