Update packages.md
Renamed "Project Subdirectories" to "Package Subdirectories". Below is an excerpt of a thread i had with @mirnawong1 
"IMO what that describes is more of a “package subdirectory” than “project subdirectory”. The confusion though is that in dbt land underlying a package is actually just a dbt project. But the use case that’s described in that documentation is a little different than what’s detailed here. 
For example, the “package subdirectory” might typically be used to seperate different functions or macros within a bigger package like macros for currency conversion might live in a different “package subdirectory” than the macros for daylight saving converstions. 
The “project subdirectory” might be to seperate different projects within a monorepo. Might be an organisational strategy to have all teams work on the same git repo but have different subdirectories for the product team or finance team."