Add token field to oauth-secrets example
I just now beat my head against the wall for half a day trying to get dbt to connect BQ following the example given here for the `oauth-secrets` method. Eventually I looked in the dbt Python code and found that the credentials are being created using the `token` field, which is in the JSON I'm pulling from but not the example in these docs. I didn't trace through the code to figure out where dbt is getting the token value into the profile_credentials dict, but I took a leap of faith and supplied the token value from my JSON in profiles.yml. And it worked immediately! So if a user follows the updated example, it should be problem solved.