slim down slim ci description
I restructured a few lines with the goal of removing a link to an anchor which was only a few lines down. (I was actually disoriented when I clicked on the link and found that the page had only scrolled down a few lines.) 

While going about this, I couldn't help making a few other changes that I think make the description of the (undeniably confusing) concept of Slim CI a bit more readable.