Switch from dots to underscores for versioning (#3316)
## What are you changing in this pull request and why?
The docs for model versioning use `dim_customers.v1` in a lot of places,
instead of `dim_customers_v1`. This was an original proposal for model
version naming that got abandoned due to database peculiarities.

This PR swaps out the `.` for `_` everywhere except for when it's being
used in a model's fully qualified name (where the dots are actually how
it's represented.)

I'm assuming this was an oversight as the product evolved; maybe it's on
purpose across the board? If so then this is more up for debate, but IMO
the more time we spend talking about versioned models along our
preferred naming scheme (single digit with an underscore) the better.
The fact that we use dots behind the scenes isn't relevant to most
folks.

Co-authored-by: Ly Nguyen <107218380+nghi-ly@users.noreply.github.com>
Co-authored-by: mirnawong1 <89008547+mirnawong1@users.noreply.github.com>