Update deploy-jobs.md for self-deferral
The CI Phase 2 beta will now include the ability for Deploy jobs to self-defer. This will be useful for jobs running compile, source_freshness:fresher, and other commands that need to defer to the last successful state of that particular job.

Feel free to suggest other language as well, this is not perfect.