Update extended attributes
Add an additional bullet to clarify that extended attributes only support top-level keys / fields. This will prevent confusion in case users try to pass in partial or individual sub-key values (instead of passing in the entire top-level key as a nested JSON block in the YAML). The latter will work while the former will cause an error when you actually try to run dbt.