Update 2023-12-15-serverless-free-tier-data-stack-with-dlt-and-dbt-core.md
Made amendments to the blog aligned with the commented feedback you provided along the lines of providing greater context. Happy to make further changes if you think deem them necessary. 

Also I imagine we will need to update the blog dates etc. aligned with when this is ready to be pushed? The dates I chose were related to the original PR.