Update ci-jobs.md
Added a suggestion on a recently reported issue with CI job run status update as per [this Zendesk ticket](https://dbtcloud.zendesk.com/agent/tickets/60166). Please let me know if you'd like the contents of the correspondence in the ticket to be shared. This has been tried, tested and confirmed internally.