Update metrics-overview.md
Hey @mirnawong1 . I learned recently that "meta" was deprecated, and instead we need to add a "meta" under config for metric, kind of like here with semantic-model.

```semantic_models:
  - name: semantic_model_name
    config:
      meta: {<dictionary>}
```

We haven't added it yet, so for now I just removed meta. Once we add it under "config", we can add an example back. cc: @Jstein77