Update dbt-commands.md
Hi.
I addedd a "()" to the "X" option to be coherent, also the "(N/A)" at the same line, for clarity. The reader shouldn't be forced to look for it's meaning as a Table note.
Lastly, I removed the last "command", since (as it's own explanation says) it's a flag, not a command. As a matter of fact, "--version" would be a good candidate for Table note.
Regards.
IL.