Update json-artifacts.md
The position of flag in the command changed: According to https://docs.getdbt.com/reference/global-configs/command-line-options the correct order is dbt run -FLAG, the option dbt -FLAG run is a legacy function and can be removed anytime. IMHO it is suboptimal to show legacy functional in the documentation instead of current one.