Fix incorrect `create_metric_display_name` reference in measure params
The measure parameters snippet contains an outdated reference to a construct named `create_metric_display_name`, which was part of the original draft spec but never got built.

Instead, we decided to call `display_name` `label`, and to automatically propagate `label` and `description` from the underlying measure to the metric we generate when `create_metric` is set to true. This update reflects that behavior.