Add section headers to separate the different sections (#5776)
[Preview](https://docs-getdbt-com-git-dbeatty10-patch-1-dbt-labs.vercel.app/reference/resource-configs/contract)

## What are you changing in this pull request and why?

The
[`contracts`](https://docs.getdbt.com/reference/resource-configs/contract#example)
page includes some important information as it relates to size,
precision, and scale when enforcing contracts, but it's hard to pick out
visually.

I want to be send hyperlinks to folks that points to just the right
section. For example, in this issue:
https://github.com/dbt-labs/dbt-core/issues/10430.

## Checklist
- [x] Review the [Content style
guide](https://github.com/dbt-labs/docs.getdbt.com/blob/current/contributing/content-style-guide.md)
so my content adheres to these guidelines.