migration: add callout for rescheduling mechanism
recently, the multicell team added a component to the in-app
migration guide which allows customers to postpone their
account migration by up to 45 days (this functionality is
currently only available to enterprise tier accounts).

as we begin migrating more enterprise accounts, we want our
public facing docs to highlight this for our users.