Update invite-users.md
I think as part of the new side nav, the steps to get to the invite users CTA may be different? I'm not an admin so I am not 100%, but I do know that I no longer see an "account settings" in my upper right corner regardless of my invite perms. Feel free to reject if I have that wrong!