Update explore-global-nav.md
I'm trying to make it clear that technically starter plans just get a new ui, not an entirely new experience. I'm going to make changes to the video as well. 

Note technically "global navigation" is a bit misleading, but I think I want to stick with it since that's the benefit for the most highest paying customers.